requirements.txt
psycopg2-binary
flask
flask-sqlalchemy
tqdm
requests
python-whois
python-dotenv